requirements.txt
flask
flask-babel
flask-login
flask-bcrypt
flask-wtf
flask-sqlalchemy
python-decouple
python-dotenv